requirements.txt
Flask==0.10.1
Flask-Session==0.3.1
watson-developer-cloud==0.26.0
python-twitter==3.3